+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "vortex-array",
     "vortex-btrblocks",
     "vortex-buffer",
-    "vortex-common",
     "vortex-datafusion",
     "vortex-dtype",
     "vortex-duckdb",
@@ -27,6 +26,7 @@ members = [
     "vortex-python",
     "vortex-scalar",
     "vortex-tui",
+    "vortex-utils",
     "xtask",
 ]
 exclude = ["duckdb-vortex/duckdb-rs", "wasm-test"]
@@ -180,7 +180,6 @@ vortex-array = { version = "0.1.0", path = "./vortex-array", default-features = 
 vortex-btrblocks = { version = "0.1.0", path = "./vortex-btrblocks", default-features = false }
 vortex-buffer = { version = "0.1.0", path = "./vortex-buffer", default-features = false }
 vortex-bytebool = { version = "0.1.0", path = "./encodings/bytebool", default-features = false }
-vortex-common = { version = "0.1.0", path = "./vortex-common", default-features = false }
 vortex-datafusion = { version = "0.1.0", path = "./vortex-datafusion", default-features = false }
 vortex-datetime-parts = { version = "0.1.0", path = "./encodings/datetime-parts", default-features = false }
 vortex-decimal-byte-parts = { version = "0.1.0", path = "encodings/decimal-byte-parts", default-features = false }
@@ -203,6 +202,7 @@ vortex-scalar = { version = "0.1.0", path = "./vortex-scalar", default-features 
 vortex-sequence = { version = "0.1.0", path = "encodings/sequence", default-features = false }
 vortex-sparse = { version = "0.1.0", path = "./encodings/sparse", default-features = false }
 vortex-tui = { version = "0.1.0", path = "./vortex-tui", default-features = false }
+vortex-utils = { version = "0.1.0", path = "./vortex-utils", default-features = false }
 vortex-zigzag = { version = "0.1.0", path = "./encodings/zigzag", default-features = false }
 # END crates published by this project
