+++ bench-vortex/src/bin/tpch.rs
@@ -28,8 +28,8 @@ use itertools::Itertools;
 use log::{info, warn};
 use similar::{ChangeTag, TextDiff};
 use url::Url;
-use vortex::common::aliases::hash_map::HashMap;
 use vortex::error::VortexExpect;
+use vortex::utils::aliases::hash_map::HashMap;
 use vortex_datafusion::persistent::metrics::VortexMetricsFinder;
 
 #[derive(Parser, Debug)]
