+++ bench-vortex/src/public_bi.rs
@@ -27,10 +27,10 @@ use tokio::runtime::Handle;
 use tracing::{debug, info};
 use url::Url;
 use vortex::ArrayRef;
-use vortex::common::aliases::hash_map::HashMap;
 use vortex::error::{VortexResult, vortex_err};
 use vortex::file::{VortexLayoutStrategy, VortexOpenOptions, VortexWriteOptions};
 use vortex::stream::ArrayStreamExt;
+use vortex::utils::aliases::hash_map::HashMap;
 use vortex_datafusion::persistent::VortexFormat;
 
 use crate::conversions::parquet_to_vortex;
