+++ bench-vortex/src/random_access/take.rs
@@ -14,9 +14,9 @@ use parquet::file::metadata::RowGroupMetaData;
 use stream::StreamExt;
 use tokio::runtime::Handle;
 use vortex::buffer::Buffer;
-use vortex::common::aliases::hash_map::HashMap;
 use vortex::file::VortexOpenOptions;
 use vortex::stream::ArrayStreamExt;
+use vortex::utils::aliases::hash_map::HashMap;
 use vortex::{Array, ArrayRef, IntoArray};
 
 pub async fn take_vortex_tokio(path: &Path, indices: Buffer<u64>) -> anyhow::Result<ArrayRef> {
