+++ encodings/dict/Cargo.toml
@@ -26,12 +26,12 @@ rand = { workspace = true, optional = true }
 rustc-hash = { workspace = true }
 vortex-array = { workspace = true }
 vortex-buffer = { workspace = true }
-vortex-common = { workspace = true }
 vortex-dtype = { workspace = true }
 vortex-error = { workspace = true }
 vortex-fsst = { workspace = true, optional = true }
 vortex-mask = { workspace = true }
 vortex-scalar = { workspace = true }
+vortex-utils = { workspace = true }
 
 [lints]
 workspace = true
