+++ encodings/dict/src/builders/primitive.rs
@@ -8,9 +8,9 @@ use vortex_array::arrays::{NativeValue, PrimitiveArray};
 use vortex_array::validity::Validity;
 use vortex_array::{Array, ArrayRef, IntoArray, ToCanonical};
 use vortex_buffer::BufferMut;
-use vortex_common::aliases::hash_map::{Entry, HashMap};
 use vortex_dtype::{NativePType, Nullability, PType};
 use vortex_error::{VortexExpect, VortexResult, vortex_bail, vortex_panic};
+use vortex_utils::aliases::hash_map::{Entry, HashMap};
 
 use super::DictConstraints;
 use crate::builders::DictEncoder;
