+++ fuzz/Cargo.toml
@@ -27,12 +27,12 @@ tokio = { workspace = true, features = ["full"] }
 vortex-array = { workspace = true, features = ["arbitrary"] }
 vortex-btrblocks = { workspace = true }
 vortex-buffer = { workspace = true }
-vortex-common = { workspace = true }
 vortex-dtype = { workspace = true, features = ["arbitrary"] }
 vortex-error = { workspace = true }
 vortex-file = { workspace = true }
 vortex-mask = { workspace = true }
 vortex-scalar = { workspace = true, features = ["arbitrary"] }
+vortex-utils = { workspace = true }
 
 [lints]
 workspace = true
