+++ fuzz/fuzz_targets/file_io.rs
@@ -11,11 +11,11 @@ use vortex_array::arrow::IntoArrowArray;
 use vortex_array::compute::{Operator, compare};
 use vortex_array::{Array, ArrayRef, Canonical, IntoArray, ToCanonical};
 use vortex_buffer::ByteBufferMut;
-use vortex_common::aliases::DefaultHashBuilder;
-use vortex_common::aliases::hash_set::HashSet;
 use vortex_dtype::{DType, StructFields};
 use vortex_error::{VortexExpect, VortexUnwrap, vortex_panic};
 use vortex_file::{VortexOpenOptions, VortexWriteOptions};
+use vortex_utils::aliases::DefaultHashBuilder;
+use vortex_utils::aliases::hash_set::HashSet;
 
 fuzz_target!(|array_data: ArbitraryArray| -> Corpus {
     let array_data = array_data.0;
