+++ vortex-array/Cargo.toml
@@ -52,7 +52,6 @@ rustc-hash = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 static_assertions = { workspace = true }
 vortex-buffer = { workspace = true, features = ["arrow"] }
-vortex-common = { workspace = true }
 vortex-dtype = { workspace = true, features = ["arrow", "serde"] }
 vortex-error = { workspace = true, features = [
     "flatbuffers",
@@ -63,6 +62,7 @@ vortex-error = { workspace = true, features = [
 vortex-flatbuffers = { workspace = true, features = ["array"] }
 vortex-mask = { workspace = true }
 vortex-scalar = { workspace = true }
+vortex-utils = { workspace = true }
 
 [features]
 arbitrary = [
