+++ vortex-array/src/arrays/primitive/top_value.rs
@@ -1,11 +1,11 @@
 use std::hash::Hash;
 
 use rustc_hash::FxBuildHasher;
-use vortex_common::aliases::hash_map::HashMap;
 use vortex_dtype::{NativePType, match_each_native_ptype};
 use vortex_error::{VortexExpect, VortexResult};
 use vortex_mask::{AllOr, Mask};
 use vortex_scalar::PValue;
+use vortex_utils::aliases::hash_map::HashMap;
 
 use crate::arrays::{NativeValue, PrimitiveArray};
