+++ vortex-btrblocks/Cargo.toml
@@ -24,7 +24,6 @@ rustc-hash = { workspace = true }
 vortex-alp = { workspace = true }
 vortex-array = { workspace = true }
 vortex-buffer = { workspace = true }
-vortex-common = { workspace = true }
 vortex-datetime-parts = { workspace = true }
 vortex-decimal-byte-parts = { workspace = true }
 vortex-dict = { workspace = true }
@@ -36,6 +35,7 @@ vortex-mask = { workspace = true }
 vortex-runend = { workspace = true }
 vortex-scalar = { workspace = true }
 vortex-sparse = { workspace = true }
+vortex-utils = { workspace = true }
 vortex-zigzag = { workspace = true }
 
 [dev-dependencies]
