+++ vortex-btrblocks/src/float/stats.rs
@@ -5,11 +5,11 @@ use num_traits::Float;
 use rustc_hash::FxBuildHasher;
 use vortex_array::ToCanonical;
 use vortex_array::arrays::{NativeValue, PrimitiveArray, PrimitiveVTable};
-use vortex_common::aliases::hash_set::HashSet;
 use vortex_dtype::half::f16;
 use vortex_dtype::{NativePType, PType};
 use vortex_error::{VortexExpect, VortexUnwrap, vortex_panic};
 use vortex_mask::AllOr;
+use vortex_utils::aliases::hash_set::HashSet;
 
 use crate::sample::sample;
 use crate::{CompressorStats, GenerateStatsOptions};
