+++ vortex-btrblocks/src/integer.rs
@@ -680,8 +680,8 @@ mod tests {
     use vortex_array::vtable::ValidityHelper;
     use vortex_array::{Array, IntoArray, ToCanonical};
     use vortex_buffer::{Buffer, BufferMut, buffer, buffer_mut};
-    use vortex_common::aliases::hash_set::HashSet;
     use vortex_sparse::SparseEncoding;
+    use vortex_utils::aliases::hash_set::HashSet;
 
     use crate::integer::{IntCompressor, IntegerStats, SparseScheme};
     use crate::{Compressor, CompressorStats, Scheme};
