+++ vortex-btrblocks/src/string.rs
@@ -1,11 +1,11 @@
 use vortex_array::arrays::{VarBinArray, VarBinViewArray, VarBinViewVTable};
 use vortex_array::vtable::ValidityHelper;
 use vortex_array::{ArrayRef, IntoArray, ToCanonical};
-use vortex_common::aliases::hash_set::HashSet;
 use vortex_dict::DictArray;
 use vortex_dict::builders::dict_encode;
 use vortex_error::{VortexExpect, VortexResult};
 use vortex_fsst::{FSSTArray, fsst_compress, fsst_train_compressor};
+use vortex_utils::aliases::hash_set::HashSet;
 
 use crate::integer::IntCompressor;
 use crate::sample::sample;
