+++ vortex-expr/src/pruning/pruning_predicate.rs
@@ -3,10 +3,10 @@ use std::iter;
 use itertools::Itertools;
 use vortex_array::stats::Stat;
 use vortex_array::{Array, ArrayRef};
-use vortex_common::aliases::hash_map::HashMap;
-use vortex_common::aliases::hash_set::HashSet;
 use vortex_dtype::{Field, FieldName, FieldPath};
 use vortex_error::{VortexExpect, VortexResult};
+use vortex_utils::aliases::hash_map::HashMap;
+use vortex_utils::aliases::hash_set::HashSet;
 
 use super::relation::Relation;
 use crate::{AccessPath, ExprRef, Scope, StatsCatalog, get_item, var};
