+++ vortex-expr/src/registry.rs
@@ -1,9 +1,9 @@
 use std::sync::LazyLock;
 
 use expr::Expr;
-use vortex_common::aliases::hash_map::HashMap;
 use vortex_error::{VortexResult, vortex_err};
 use vortex_proto::expr;
+use vortex_utils::aliases::hash_map::HashMap;
 
 use crate::between::proto::BetweenSerde;
 use crate::binary::proto::BinarySerde;
