+++ vortex-expr/src/scope.rs
@@ -3,9 +3,9 @@ use std::str::FromStr;
 use std::sync::Arc;
 
 use vortex_array::{Array, ArrayRef};
-use vortex_common::aliases::hash_map::HashMap;
 use vortex_dtype::DType;
 use vortex_error::{VortexError, VortexResult, vortex_bail, vortex_err};
+use vortex_utils::aliases::hash_map::HashMap;
 
 type ExprScope<T> = HashMap<Identifier, T>;
