+++ vortex-file/src/driver.rs
@@ -9,11 +9,11 @@ use futures::{Stream, StreamExt};
 use itertools::Itertools;
 use linked_hash_set::LinkedHashSet;
 use vortex_buffer::{Alignment, ByteBuffer};
-use vortex_common::aliases::hash_map::HashMap;
 use vortex_error::{VortexError, VortexExpect, VortexResult, vortex_panic};
 use vortex_io::{PerformanceHint, VortexReadAt};
 use vortex_layout::segments::{SegmentEvent, SegmentId, SegmentRequest};
 use vortex_metrics::{Counter, VortexMetrics};
+use vortex_utils::aliases::hash_map::HashMap;
 
 use crate::SegmentSpec;
