+++ vortex-layout/src/sequence.rs
@@ -7,8 +7,8 @@ use std::sync::Arc;
 use std::task::{Context, Poll, Waker};
 
 use parking_lot::Mutex;
-use vortex_common::aliases::hash_map::HashMap;
 use vortex_error::VortexExpect;
+use vortex_utils::aliases::hash_map::HashMap;
 
 use crate::segments::SegmentId;
